pyproject.toml
[tool.poetry]
name = "txt-visual-seperator"
version = "0.1.0"
description = "a program that separates text and visual elements from image."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-vision = "^3.7.2"
python-dotenv = "^1.0.1"
opencv-python = "^4.9.0.80"
ultralytics = "^8.2.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
